added a few try/catches to NTR connections so if it fails it will still connect to IR and discord. I personally use this for a friend code mode on old 3ds that have issues with NTR and the friendlist. updated auto legality mod and pkhex deps updated pk file reads to display correct trainer info per generation.

+++ discord/TradeModule.cs
@@ -75,7 +75,7 @@ namespace _3DS_link_trade_bot
                 }
 
 
-                if (Legal.ZCrystalDictionary.ContainsValue(pkm.HeldItem))
+                if (Legal.ZCrystalDictionary.ContainsValue((ushort)pkm.HeldItem))
                     pkm.HeldItem = 0;
                 if (!new LegalityAnalysis(pkm).Valid || FormInfo.IsFusedForm(pkm.Species,pkm.Form,pkm.Format))
                 {
